rbac.session: Implement RBAC session without database persistence
Previously all RBAC sessions were represented by `models.RbacSession` and
are stored in the database. This new approach does not store the RBAC
sessions in the database any more. For backend sessions (non-web sessions)
this never made sense in the first place.

This has multiple benefits:

- We don't need to invalidate RBAC sessions after changing roles any more
- We don't need to run regular cleanups any more
- We don't fill the database with mostly useless information

RBAC sessions in the web context could still be kept persistent for a user
by simply saving the session user's RBAC roles in the Django session.